GT-608 Fix termination issue in Process tests
