+++ test/docker_restart_noagent_member_test.go
@@ -140,11 +140,3 @@ func TestDockerRestartNoAgentMember(t *testing.T) {
 	waitForCallFunction(t, getShutdownCalls(members)...)
 	removeDockerVolumesByLabel(t, "starter-test=true")
 }
-
-func getShutdownCalls(members map[int]MembersConfig) []callFunction {
-	var calls []callFunction
-	for _, m := range members {
-		calls = append(calls, ShutdownStarterCall(fmt.Sprintf("http://localhost:%d", m.Port)))
-	}
-	return calls
-}
