+++ test/process_restart_noagent_member_test.go
@@ -117,51 +117,3 @@ func TestProcessMultipleRestartNoAgentMember(t *testing.T) {
 
 	waitForCallFunction(t, getShutdownCalls(members)...)
 }
-
-func TestProcessAgentsMultipleRestart(t *testing.T) {
-	removeArangodProcesses(t)
-	testMatch(t, testModeProcess, starterModeCluster, false)
-
-	members := map[int]MembersConfig{
-		6000:  {"node1", 6000, SetUniqueDataDir(t), nil, nil},
-		7000:  {"node2", 7000, SetUniqueDataDir(t), nil, nil},
-		8000:  {"node3", 8000, SetUniqueDataDir(t), nil, nil},
-		9000:  {"node4", 9000, SetUniqueDataDir(t), nil, nil},
-		10000: {"node5", 10000, SetUniqueDataDir(t), nil, nil},
-	}
-
-	joins := "localhost:6000,localhost:7000,localhost:8000"
-	for port, m := range members {
-		m.Process = spawnMemberProcess(t, m.Port, m.DataDir, joins, "")
-		members[port] = m
-	}
-
-	waitForCluster(t, members, time.Now())
-
-	t.Logf("Verify setup.json after fresh start")
-	verifyProcessSetupJson(t, members, 3)
-	verifyEndpointSetup(t, members)
-
-	for i := 0; i < 1; i++ {
-		t.Logf("Restart all members, iteration: %d", i)
-		t.Run("Restart all members", func(t *testing.T) {
-			for port := range members {
-				require.NoError(t, members[port].Process.Kill())
-			}
-			time.Sleep(3 * time.Second)
-
-			for port, m := range members {
-				m.Process = spawnMemberProcess(t, m.Port, m.DataDir, joins, "")
-				members[port] = m
-			}
-
-			waitForCluster(t, members, time.Now())
-
-			t.Logf("Verify setup after member restart, iteration: %d", i)
-			verifyProcessSetupJson(t, members, 3)
-			verifyEndpointSetup(t, members)
-		})
-	}
-
-	waitForCallFunction(t, getShutdownCalls(members)...)
-}
